gaussian_splatting: share one StreamingTierCapPolicy; drop duplicate + local settings re-impls
The StreamingTierCapPolicy struct and its resolver functions
(_project_setting_has_override, _resolve_tiered_cap_uint,
_resolve_streaming_tier_cap_policy) were duplicated near-verbatim in
streaming_upload_pipeline.cpp and streaming_vram_regulator.cpp — identical
logic, differing only in that the upload pipeline already read settings via
the canonical gs::settings::get_* while the regulator used three local
re-implementations (_ps_get_uint/_ps_get_bool/_ps_get_float) that themselves
just duplicated gs::settings.

Consolidate into a single shared header core/streaming_tier_cap_policy.h
(namespace gs_tier_cap, inline resolvers) that reads settings only through
gs::settings. Both .cpp now include it + add `using namespace gs_tier_cap;`
so call sites are unchanged. With the resolvers shared, the regulator's local
_ps_get_* helpers became redundant: all five direct call sites were migrated
to gs::settings::* and the helpers deleted.

Net ~173 LOC of duplication removed for ~90 LOC of shared header. Behavior-
preserving (gs::settings::get_* matches the prior getters for the BOOL/UINT/
FLOAT-typed settings involved; the upload pipeline already used this path).

Note: a third, struct-only StreamingTierCapPolicy copy in gaussian_streaming.cpp
remains (no resolvers, separate anon namespace, does not include this header so
no collision) — left for a future de-dup pass.

Risk: R2 (streaming). Evidence: full editor build compiles + links clean
(scons ... dev_build=yes tests=yes, exit 0; only pre-existing C4458 warnings in
gpu_buffer_manager.h); guards green (run_module_tests.py --guard-only, 108 OK).
